cvicenie 6

+++ src/app/user/user-page/user-page.component.ts
@@ -2,6 +2,8 @@ import {Component} from '@angular/core';
 import {User} from '../../common/model/user.model';
 import {UserService} from '../../common/service/user.service';
 import {UntilDestroy, untilDestroyed} from '@ngneat/until-destroy';
+import {ToastService} from 'angular-toastify';
+import {Router} from '@angular/router';
 
 @UntilDestroy()
 @Component({
@@ -13,9 +15,9 @@ export class UserPageComponent {
 
   persons: Array<User> = [];
 
-  person?: User;
-
-  constructor(private service: UserService) {
+  constructor(private service: UserService,
+              private toastService: ToastService,
+              private router: Router) {
     this.getPersons();
   }
 
@@ -27,28 +29,25 @@ export class UserPageComponent {
 
   createPerson(person: User): void {
     this.service.createUser(person).pipe(untilDestroyed(this)).subscribe(() => {
-      console.log('Osoba bola úspešne uložená.');
+      this.toastService.success('Osoba bola úspešne uložená.');
       this.getPersons();
-    })
-  }
-
-  updatePerson(person: User): void {
-    this.service.updateUser(person).pipe(untilDestroyed(this)).subscribe(() => {
-      console.log('Osoba bola úspešne zmenená.');
-      this.getPersons();
+    }, () => {
+      this.toastService.error('Chyba. Osoba nebola uložená.');
     })
   }
 
   selectPersonToUpdate(personId: number): void {
-    this.service.getUser(personId).pipe(untilDestroyed(this)).subscribe((person: User) => {
-      this.person = person;
-    });
+    this.router.navigate(['user', personId]);
   }
 
   deletePerson(personId: number): void {
-    this.service.deleteUser(personId).pipe(untilDestroyed(this)).subscribe(() => {
-      console.log('Osoba bola úspešne zmazaná.');
-      this.getPersons();
-    })
+    if (window.confirm('Naozaj chcete vymazať osobu?')) {
+      this.service.deleteUser(personId).pipe(untilDestroyed(this)).subscribe(() => {
+        this.toastService.success('Osoba bola úspešne zmazaná.');
+        this.getPersons();
+      }, () => {
+        this.toastService.error('Chyba. Osoba nebola zmazaná.');
+      })
+    }
   }
 }
